feat(Experimental): Add "rootInjector", a partner to "rootLegacyInjector".
PiperOrigin-RevId: 258656845
